+++ rdf-export/rdf-export-application/src/test/kotlin/eu/tib/orkg/prototype/export/rdf/domain/StatementRDFTest.kt
@@ -9,7 +9,6 @@ import eu.tib.orkg.prototype.statements.api.Classes
 import eu.tib.orkg.prototype.statements.api.Predicates
 import eu.tib.orkg.prototype.statements.domain.model.ThingId
 import org.assertj.core.api.Assertions.assertThat
-import org.junit.jupiter.api.Disabled
 import org.junit.jupiter.api.Test
 
 @Suppress("HttpUrlsUsage")
@@ -43,7 +42,20 @@ internal class StatementRDFTest {
     }
 
     @Test
-    @Disabled("We need the full URI. This might need a change in the API and needs to be discussed.")
+    fun `converts statement with literal object and XSD datatype to NTriple format correctly`() {
+        val expectedOutput = """
+            |<http://orkg.org/orkg/resource/R100> <http://orkg.org/orkg/predicate/P200> "object"^^<http://www.w3.org/2001/XMLSchema#integer> .
+            |
+        """.trimMargin()
+        val statement = createStatement(
+            subject = createResource().copy(id = ThingId("R100"), label = "subject"),
+            predicate = createPredicate().copy(id = ThingId("P200")),
+            `object` = createLiteral().copy(id = ThingId("L300"), label = "object", datatype = "xsd:integer")
+        )
+        assertThat(statement::toNTriple.asString()).isEqualTo(expectedOutput)
+    }
+
+    @Test
     fun `converts statement with literal object and custom datatype to NTriple format correctly`() {
         val expectedOutput = """
             |<http://orkg.org/orkg/resource/R100> <http://orkg.org/orkg/predicate/P200> "object"^^<http://example.org/myDataType> .